+++ scripts/src/admin/ResponseTable/ResponseTable.tsx
@@ -9,32 +9,40 @@ import { fetchResponses, setResponsesSelectedView } from '../../store/responses/
 import { IResponseTableProps, IResponseTableState } from "./ResponseTable.d";
 import ResponseTableView from "./ResponseTableView";
 import "./ResponseTable.scss";
+import { push } from 'connected-react-router';
 
 
 class ResponseTable extends React.Component<IResponseTableProps, IResponseTableState> {
-    static defaultProps = {
-        editMode: false,
-        checkinMode: false
-    }
 
     componentDidMount() {
         this.props.fetchRenderedForm(this.props.match.params.formId).then(() => this.props.fetchResponses(this.props.match.params.formId));
     }
 
     render() {
-        return (!this.props.responses || !this.props.form) ? <Loading /> : <ResponseTableView />;
+        if (!this.props.responses || !this.props.form) {
+            return <Loading />;
+        }
+        return (
+            <ResponseTableView
+                responses={this.props.responses}
+                renderedForm={this.props.form.renderedForm}
+                tableViewName={this.props.tableViewName}
+                push={(e) => this.props.push(e)}
+                />);
     }
 }
 
 const mapStateToProps = state => ({
     ...state.responses,
-    form: state.form
+    form: state.form,
+    tableViewName: (state.router.location.pathname.match(/\/(.[a-zA-Z_]*?)$/) || [null, null])[1]
 });
 
 const mapDispatchToProps = (dispatch) => ({
     fetchResponses: formId => dispatch(fetchResponses(formId)),
     fetchRenderedForm: formId => dispatch(fetchRenderedForm(formId)),
-    setResponsesSelectedView: (e: string) => dispatch(setResponsesSelectedView(e))
+    setResponsesSelectedView: (e: string) => dispatch(setResponsesSelectedView(e)),
+    push: (e: string) => dispatch(push(`./${e}`))
 });
 
 export default connect(mapStateToProps, mapDispatchToProps)(ResponseTable);